ClientApp/src/app/task/task.component.ts
import { Component, OnInit, Input, Output, EventEmitter } from '@angular/core';
import { Task } from '../task';
import { TasksService } from '../tasks.service';

@Component({
  selector: 'app-task',
  templateUrl: './task.component.html'
})
export class TaskComponent implements OnInit {
  @Input() task: Task;
  @Output() change: EventEmitter<void> = new EventEmitter<void>();

  constructor(
    private tasksService: TasksService
  ) { }

  ngOnInit() {
  }

  nextState() {
    this.task.state++;
    this.modifyAndNotify();
  }

  prevState() {
    this.task.state--;
    this.modifyAndNotify();
  }

  remove() {
    this.tasksService.delete(this.task.id)
      .subscribe(() => {
        this.change.emit();
      });
  }

  private modifyAndNotify() {
    this.tasksService.modify(this.task)
      .subscribe(() => {
        this.change.emit();
      });
  }

}
